Transform, access, and part of the documentation
Large commit. It includes:
Two new functions - acled_transform and acled_api - with their documentations.

Some small changes to the acled_api to account for acled_access (currently an opt-out), and some warning messages (although this is still on the works).

Edits of some of the documentation. Adding documentation for the two new functions, plus a vignitte for acled_transform.

After this commit, I will be doing smaller commits for better traceability.