Delete transform & acled_access & new api
In this commit I took out the transform function per request of Rachel.

I modified the acled_access so that it runs a check of the credentials first, and stores it later.

I modified the acled_api function so that now it includes timestamp, and extra.

Address the new workflow in the vignitte acled_api vignitte, including more examples for the new arguments.

I also added a to_do list so that we can track what we are changing in every commit.

I made some changes to the methodology parts, but take these with a grain of salt, because we may remove all of them soon (I want to confirm that).